+++ src/ethereum/info/botanixInfo.ts
@@ -16,17 +16,9 @@ import {
   makeEvmDefaultSettings
 } from './ethereumCommonInfo'
 
-const builtinTokens: EdgeTokenMap = {
-  '94373a4919b3240d86ea41593d5eba789fef3848': {
-    currencyCode: 'WETH',
-    displayName: 'Wrapped Ethereum',
-    denominations: [{ name: 'WETH', multiplier: '1000000000000000000' }],
-    networkLocation: {
-      contractAddress: '0x94373a4919B3240D86eA41593D5eBa789FEF3848'
-    }
-  }
-}
+const builtinTokens: EdgeTokenMap = {}
 
+// Fees are in Wei
 const networkFees: EthereumFees = {
   default: {
     baseFee: undefined,
@@ -48,86 +40,71 @@ const networkFees: EthereumFees = {
       standardFeeLowAmount: '100000000000000000',
       standardFeeHighAmount: '10000000000000000000',
       highFee: '40000000001',
-      minGasPrice: '1000000000'
+      minGasPrice: '100000'
     },
     minPriorityFee: '2000000000'
   }
 }
 
 const networkInfo: EthereumNetworkInfo = {
-  addressQueryLookbackBlocks: 9, // ~2 minutes
+  addressQueryLookbackBlocks: 36, // 2 minutes
   networkAdapterConfigs: [
     {
       type: 'rpc',
-      servers: [
-        'https://holesky-fullnode-testnet.rpc.grove.city/v1/{{poktPortalApiKey}}',
-        'https://eth-holesky.alchemyapi.io/v2/{{alchemyApiKey}}',
-        'https://holesky.infura.io/v3/{{infuraProjectId}}',
-        'https://1rpc.io/holesky',
-        'https://ethereum-holesky-rpc.publicnode.com',
-        'https://ethereum-holesky.blockpi.network/v1/rpc/public',
-        'https://holesky.drpc.org',
-        'https://rpc-holesky.rockx.com',
-        'https://endpoints.omniatech.io/v1/eth/holesky/public',
-        'https://lb.drpc.org/ogrpc?network=holesky&dkey={{drpcApiKey}}'
-      ],
-      ethBalCheckerContract: '0xf4C055E8760C6e66301C2e2F12b85567a9A50841'
+      servers: ['https://rpc.botanixlabs.com']
     },
     {
       type: 'evmscan',
-      servers: ['https://api.etherscan.io']
+      servers: [
+        'https://api.routescan.io/v2/network/mainnet/evm/3637/etherscan'
+      ]
     }
   ],
-
-  uriNetworks: ['ethereum', 'ether'],
+  uriNetworks: [],
   ercTokenStandard: 'ERC20',
   chainParams: {
-    chainId: 17000,
-    name: 'Holesky'
+    chainId: 3637,
+    name: 'BTC'
   },
+  optimismRollup: false,
   supportsEIP1559: true,
   hdPathCoinType: 60,
-  pluginMnemonicKeyName: 'holeskyMnemonic',
-  pluginRegularKeyName: 'holeskyKey',
+  pluginMnemonicKeyName: 'botanixMnemonic',
+  pluginRegularKeyName: 'botanixKey',
   evmGasStationUrl: null,
   networkFees
 }
 
 const currencyInfo: EdgeCurrencyInfo = {
   canReplaceByFee: true,
-  currencyCode: 'ETH',
+  currencyCode: 'BTC',
   customFeeTemplate: evmCustomFeeTemplate,
   customTokenTemplate: evmCustomTokenTemplate,
-  chainDisplayName: 'Holesky Testnet',
-  assetDisplayName: 'Holesky Testnet',
+  assetDisplayName: 'BTC',
+  chainDisplayName: 'BTC',
   memoOptions: evmMemoOptions,
-  pluginId: 'holesky',
-  walletType: 'wallet:holesky',
+  pluginId: 'botanix',
+  walletType: 'wallet:botanix',
 
   // Explorers:
-  addressExplorer: 'https://holesky.etherscan.io/address/%s',
-  transactionExplorer: 'https://holesky.etherscan.io/tx/%s',
+  addressExplorer: 'https://botanixscan.io/address/%s',
+  transactionExplorer: 'https://botanixscan.io/tx/%s',
 
   denominations: [
     {
-      name: 'ETH',
+      name: 'BTC',
       multiplier: '1000000000000000000',
-      symbol: 'Ξ'
-    },
-    {
-      name: 'mETH',
-      multiplier: '1000000000000000',
-      symbol: 'mΞ'
+      symbol: '₿'
     }
   ],
 
   // Deprecated:
   defaultSettings: makeEvmDefaultSettings(networkInfo),
-  displayName: 'Holesky Testnet',
+  displayName: 'BTC',
   metaTokens: makeMetaTokens(builtinTokens)
 }
 
-export const holesky = makeOuterPlugin<
+export const botanix = makeOuterPlugin<
   EthereumNetworkInfo,
   EthereumTools,
   EthereumInfoPayload