+++ src/ethereum/info/filecoinFevmInfo.ts
@@ -81,7 +81,6 @@ const networkInfo: EthereumNetworkInfo = {
   hdPathCoinType: 461,
   feeUpdateFrequencyMs: 20000,
   supportsEIP1559: true,
-  amberDataBlockchainId: '', // ETH mainnet
   pluginMnemonicKeyName: 'filecoinfevmMnemonic',
   pluginRegularKeyName: 'filecoinfevmKey',
   evmGasStationUrl: null,