+++ src/ethereum/info/polygonInfo.ts
@@ -208,7 +208,6 @@ const networkInfo: EthereumNetworkInfo = {
   },
   supportsEIP1559: true,
   hdPathCoinType: 60,
-  amberDataBlockchainId: '',
   pluginMnemonicKeyName: 'polygonMnemonic',
   pluginRegularKeyName: 'polygonKey',
   evmGasStationUrl: 'https://gasstation.polygon.technology/v2',